requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
anyio==3.6.2
    # via httpcore
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
babel==2.11.0
    # via -r requirements.in
bcrypt==4.0.1
    # via -r requirements.in
billiard==3.6.4.0
    # via celery
bleach==5.0.1
    # via -r requirements.in
boto3==1.26.2
    # via -r requirements.in
botocore==1.29.2
    # via
    #   boto3
    #   s3transfer
celery==5.2.7
    # via -r requirements.in
certifi==2022.9.24
    # via
    #   django-healthchecks
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via
    #   cmarkgfm
    #   cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cmarkgfm==2022.10.27
    # via -r requirements.in
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==38.0.3
    # via pyjwt
decorator==5.1.1
    # via retry
deprecated==1.2.13
    # via redis
dj-rest-auth==2.2.5
    # via -r requirements.in
django==3.2.16
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-anymail
    #   django-axes
    #   django-bulk-update
    #   django-cachalot
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-hattori
    #   django-healthchecks
    #   django-markdownx
    #   django-phonenumber-field
    #   django-prettyjson
    #   django-redis
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-jwt
    #   drf-recaptcha
    #   drf-yasg
django-anymail==8.6
    # via -r requirements.in
django-axes==5.39.0
    # via -r requirements.in
django-behaviors==0.5.1
    # via -r requirements.in
django-bulk-update==2.2.0
    # via
    #   -r requirements.in
    #   django-hattori
django-cachalot==2.5.2
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-countries==7.4.2
    # via -r requirements.in
django-debug-toolbar==3.7.0
    # via -r requirements.in
django-environ==0.9.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-hattori==0.2.1
    # via -r requirements.in
django-healthchecks==1.5.0
    # via -r requirements.in
django-ipware==4.0.2
    # via
    #   -r requirements.in
    #   django-axes
    #   drf-recaptcha
django-markdownx==3.0.1
    # via -r requirements.in
django-phonenumber-field==7.0.0
    # via -r requirements.in
django-phonenumbers==1.0.1
    # via -r requirements.in
django-prettyjson==0.4.1
    # via -r requirements.in
django-redis==5.2.0
    # via -r requirements.in
django-split-settings==1.2.0
    # via -r requirements.in
django-storages==1.13.1
    # via -r requirements.in
django-tree-queries==0.11.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-jwt
    #   drf-recaptcha
    #   drf-yasg
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
drf-jwt==1.19.2
    # via -r requirements.in
drf-recaptcha==2.2.2
    # via -r requirements.in
drf-yasg==1.21.4
    # via -r requirements.in
faker==4.0.0
    # via
    #   -r requirements.in
    #   django-hattori
h11==0.12.0
    # via httpcore
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
inflection==0.5.1
    # via
    #   -r requirements.in
    #   drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.2.4
    # via celery
markdown==3.4.1
    # via django-markdownx
markupsafe==2.1.1
    # via jinja2
packaging==21.3
    # via
    #   drf-yasg
    #   redis
phonenumbers==8.13.0
    # via django-phonenumbers
pillow==9.3.0
    # via django-markdownx
prompt-toolkit==3.0.32
    # via click-repl
psycopg2-binary==2.9.5
    # via -r requirements.in
py==1.11.0
    # via retry
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-jwt
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via
    #   botocore
    #   faker
pytz==2022.6
    # via
    #   babel
    #   celery
    #   django
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0
    # via -r requirements.in
redis==4.3.4
    # via
    #   -r requirements.in
    #   django-redis
requests==2.28.1
    # via
    #   coreapi
    #   django-anymail
    #   django-healthchecks
retry==0.9.2
    # via -r requirements.in
rfc3986[idna2008]==1.5.0
    # via httpx
ruamel-yaml==0.17.21
    # via drf-yasg
s3transfer==0.6.0
    # via boto3
sentry-sdk==1.10.1
    # via -r requirements.in
shortuuid==1.0.9
    # via -r requirements.in
simplejson==3.17.6
    # via -r requirements.in
six==1.16.0
    # via
    #   bleach
    #   click-repl
    #   django-hattori
    #   django-prettyjson
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
standardjson==0.3.1
    # via django-prettyjson
text-unidecode==1.3
    # via faker
tqdm==4.64.1
    # via django-hattori
typing-extensions==4.4.0
    # via django-countries
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.12
    # via
    #   -r requirements.in
    #   botocore
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
whitenoise==6.2.0
    # via -r requirements.in
wrapt==1.14.1
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
